# Improve performance of UR calculations