--- a/osu.Game/Screens/Play/HUD/UnstableRateCounter.cs
+++ b/osu.Game/Screens/Play/HUD/UnstableRateCounter.cs
@@ -28,6 +28,8 @@ namespace osu.Game.Screens.Play.HUD
         private const float alpha_when_invalid = 0.3f;
         private readonly Bindable<bool> valid = new Bindable<bool>();
 
+        private HitEventExtensions.UnstableRateCalculationResult? unstableRateResult;
+
         [Resolved]
         private ScoreProcessor scoreProcessor { get; set; } = null!;
 
@@ -53,13 +55,20 @@ namespace osu.Game.Screens.Play.HUD
             updateDisplay();
         }
 
-        private void updateDisplay(JudgementResult _) => Scheduler.AddOnce(updateDisplay);
+        private void updateDisplay(JudgementResult result)
+        {
+            if (HitEventExtensions.AffectsUnstableRate(result.HitObject, result.Type))
+                Scheduler.AddOnce(updateDisplay);
+        }
 
         private void updateDisplay()
         {
-            double? unstableRate = scoreProcessor.HitEvents.CalculateUnstableRate()?.Result;
+            unstableRateResult = scoreProcessor.HitEvents.CalculateUnstableRate(unstableRateResult);
+
+            double? unstableRate = unstableRateResult?.Result;
 
             valid.Value = unstableRate != null;
+
             if (unstableRate != null)
                 Current.Value = (int)Math.Round(unstableRate.Value);
         }