Closes https://github.com/ppy/osu/issues/30635.

Baseline:

| Method                | Mean     | Error    | StdDev   | Allocated |
|---------------------- |---------:|---------:|---------:|----------:|
| CalculateUnstableRate | 15.15 us | 0.123 us | 0.115 us |      96 B |

With patch applied (605fe71f46eaf2944b659389cca3cfa03011e5de):

| Method                | Mean     | Error    | StdDev   | Allocated |
|---------------------- |---------:|---------:|---------:|----------:|
| CalculateUnstableRate | 11.30 us | 0.064 us | 0.060 us |      96 B |

Didn't show as much of an improvement as I expected to see based on the profiler results (seems like profiler is getting it wrong...)

With list `for` iteration (33d725e889481843601fe1f647a88ba866a8c6ce with `foreach` switched to `for`):

| Method                | Mean     | Error     | StdDev    | Allocated |
|---------------------- |---------:|----------:|----------:|----------:|
| CalculateUnstableRate | 9.917 us | 0.1824 us | 0.1706 us |      48 B |

Full PR:

| Method                                           | Mean         | Error      | StdDev     | Allocated |
|------------------------------------------------- |-------------:|-----------:|-----------:|----------:|
| CalculateUnstableRate                            | 25,305.35 us | 417.166 us | 348.352 us |   81943 B |
| CalculateUnstableRateUsingIncrementalCalculation |     24.46 us |   0.483 us |   0.537 us |      40 B |

I'm not immediately sure why it's benchmarking slower. Spent my allocated time on this so I'm going just going to smile and nod, but interested if anyone has an idea.

Importantly though, real world performance is much better (over a single autoplay run of [this beatmap](https://osu.ppy.sh/beatmapsets/367631#osu/958098)):

Baseline:

![JetBrains Rider 2024-11-25 at 12 00 17](https://github.com/user-attachments/assets/8b9b3841-c2b2-428f-a424-6c07cd84b341)

On ea68d4b33abbd920e267001fb4785ae000031f54:

![JetBrains Rider 2024-11-25 at 11 58 27](https://github.com/user-attachments/assets/5c9b4ef3-5fd6-4d6c-ab29-4526e4317602)

On bbe8f2ec44cf7f2fa76c23dddbfcc33bea7045ee (kinda OP because slider ticks and what not will not recalc at all):

![JetBrains Rider 2024-11-25 at 12 03 36](https://github.com/user-attachments/assets/22db11de-49b5-48a0-a20c-d908a8c8e799)