@peppy 2024-11-26
> cursory review because description gives impression of not being 100% set in stone yet

originally i wasn't going to PR, so it probably reads a bit that way. the stand-alone benchmark wasn't giving great improvements.

i'll take a second look at the benchmark today to see if i'm doing anything obviously wrong.